backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
psycopg[binary]==3.2.9
alembic==1.12.0
APScheduler==3.10.4
SQLAlchemy==2.0.43
